requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# Streamlit UI
streamlit==1.32.0

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
pypdf==4.0.2

# Vector Store & Embeddings
langchain==0.1.12
langchain-community==0.0.28
langchain-google-genai==0.0.9
faiss-cpu==1.8.0

# LLM
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.1
pydantic==2.6.3
numpy==1.26.4
